test: share error taxonomy harness
