+++ packages/vnf-wallet-sdk-nodejs/src/impl/data/repositories/PresentationRequestRepositoryImpl.ts
@@ -8,9 +8,7 @@ import { HttpMethod } from '../infrastructure/network/HttpMethod';
 import {
     toClientRequestFetchError,
     ErrorTaxonomy,
-    invalidLink,
 } from '../../utils/ErrorTaxonomy';
-import VelocityDeepLinkValidator from '../../utils/VelocityDeepLinkValidator';
 
 export default class PresentationRequestRepositoryImpl implements PresentationRequestRepository {
     constructor(private readonly networkService: NetworkService) {}
@@ -19,15 +17,6 @@ export default class PresentationRequestRepositoryImpl implements PresentationRe
         presentationRequestDescriptor: VCLPresentationRequestDescriptor,
     ): Promise<string> {
         const { endpoint } = presentationRequestDescriptor;
-        if (!endpoint) {
-            throw invalidLink({
-                message: 'presentationRequestDescriptor.endpoint = null',
-                sourceErrorCode:
-                    VelocityDeepLinkValidator.SourceInvalidOrMissingRequestEndpoint,
-                requestKind: ErrorTaxonomy.RequestKindPresentation,
-            });
-        }
-
         let presentationRequestResponse;
         try {
             presentationRequestResponse = await this.networkService.sendRequest(
