+++ packages/vnf-wallet-sdk-nodejs/test/infrastructure/resources/valid/PresentationSubmissionMocks.ts
@@ -5,6 +5,7 @@ import { JwtMocks } from './JwtMocks';
 import VCLVerifiableCredential from '../../../../src/api/entities/VCLVerifiableCredential';
 import VCLVerifiedProfile from '../../../../src/api/entities/VCLVerifiedProfile';
 import { DidJwkMocks } from './DidJwkMocks';
+import TokenMocks from './TokenMocks';
 
 class PresentationSubmissionMocks {
     static PushDelegate = new VCLPushDelegate(
@@ -12,11 +13,8 @@ class PresentationSubmissionMocks {
         'if0123asd129smw321',
     );
 
-    static PresentationSubmissionResultStr =
-        '{"token":"u7yLD8KS2eTEqkg9aRQE","exchange":{"id":"64131231","type":"DISCLOSURE","disclosureComplete":true,"exchangeComplete":true}}';
-
     static PresentationSubmissionResultJson = {
-        token: 'u7yLD8KS2eTEqkg9aRQE',
+        token: TokenMocks.TokenStr1,
         exchange: {
             id: '64131231',
             type: 'DISCLOSURE',
@@ -24,6 +22,10 @@ class PresentationSubmissionMocks {
             exchangeComplete: true,
         },
     };
+
+    static PresentationSubmissionResultStr = JSON.stringify(
+        PresentationSubmissionMocks.PresentationSubmissionResultJson,
+    );
 
     static PresentationRequest = new VCLPresentationRequest(
         JwtMocks.JWT,
@@ -42,6 +44,7 @@ class PresentationSubmissionMocks {
     ];
 
     static JwtEncodedSubmission =
+        // eslint-disable-next-line max-len
         'eyJraWQiOiJkaWQ6andrOmV5SnJkSGtpT2lKRlF5SXNJblZ6WlNJNkluTnBaeUlzSW1OeWRpSTZJbEF0TWpVMklpd2lhMmxrSWpvaU5qRXhZVGs1TmprdFlqWmpNeTAwWXpJeExXSTFZell0T0dRd01qa3dZMlkwTlRabElpd2llQ0k2SWtacVVsOHhVelZ2WVRkblIydElSRlJsYkhKVGIydHJjMjF0ZEZSa1VFNXRkbTFUVWtSc1FtZHBXVlVpTENKNUlqb2lTRWs0T0hOTFZ6UndlbmxWU0ZSMFFsaGtPR3RQYVZwemVGZGhUSGhxUkVWZlEwdEhWR05GVkRCUU1DSjkjMCIsInR5cCI6IkpXVCIsImFsZyI6IkVTMjU2IiwiandrIjp7Imt0eSI6IkVDIiwidXNlIjoic2lnIiwiY3J2IjoiUC0yNTYiLCJraWQiOiI2MTFhOTk2OS1iNmMzLTRjMjEtYjVjNi04ZDAyOTBjZjQ1NmUiLCJ4IjoiRmpSXzFTNW9hN2dHa0hEVGVsclNva2tzbW10VGRQTm12bVNSRGxCZ2lZVSIsInkiOiJISTg4c0tXNHB6eVVIVHRCWGQ4a09pWnN4V2FMeGpERV9DS0dUY0VUMFAwIn19.eyJzdWIiOiJlSWZXUmhqSkNMIiwibmJmIjoxNzE2NDUwMDExLCJpc3MiOiJkaWQ6andrOmV5SnJkSGtpT2lKRlF5SXNJblZ6WlNJNkluTnBaeUlzSW1OeWRpSTZJbEF0TWpVMklpd2lhMmxrSWpvaU5qRXhZVGs1TmprdFlqWmpNeTAwWXpJeExXSTFZell0T0dRd01qa3dZMlkwTlRabElpd2llQ0k2SWtacVVsOHhVelZ2WVRkblIydElSRlJsYkhKVGIydHJjMjF0ZEZSa1VFNXRkbTFUVWtSc1FtZHBXVlVpTENKNUlqb2lTRWs0T0hOTFZ6UndlbmxWU0ZSMFFsaGtPR3RQYVZwemVGZGhUSGhxUkVWZlEwdEhWR05GVkRCUU1DSjkiLCJ2cCI6eyJ0eXBlIjoiVmVyaWZpYWJsZVByZXNlbnRhdGlvbiIsInByZXNlbnRhdGlvbl9zdWJtaXNzaW9uIjp7ImlkIjoiN2U0NjM3MzktZGRlMC00OTU2LWFlNTYtMjk4MDQ5YmY3MzVlIiwiZGVmaW5pdGlvbl9pZCI6IiIsImRlc2NyaXB0b3JfbWFwIjpbXX0sInZlcmlmaWFibGVDcmVkZW50aWFsIjpbXX0sImV4cCI6MTcxNzA1NDgxMSwiaWF0IjoxNzE2NDUwMDExLCJqdGkiOiIxOTllOWRiMC1mNDg5LTQ1MjAtOGYyNy1hYjM4N2NjMzQ4MWQifQ';
 }
